+++ src/main/java/com/code/pojo/ClazzTeacher.java
@@ -24,7 +24,7 @@ public class ClazzTeacher implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "teacher_id", type = IdType.AUTO)
-    private Integer teacherId;
+    private long teacherId;
 
     private String clazzNo;
