+++ src/main/java/com/code/pojo/TStudent.java
@@ -24,7 +24,7 @@ public class TStudent implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "student_id", type = IdType.AUTO)
-    private Integer studentId;
+    private long studentId;
 
     private String studentNo;
